* Win64 port: Patches for wxSmith.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@9021 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/wxSmith/properties/wxsproperty.h
@@ -298,8 +298,8 @@ class wxsProperty
  * \param Variable name of variable inside class
  */
 #define wxsOFFSET(Class,Variable) \
-    ( (long)(((char*)(&(((Class*)1)->Variable)))) - \
-      (long)(((char*)((wxsPropertyContainer*)((Class*)1)))) )
+    ( (intptr_t)(((char*)(&(((Class*)1)->Variable)))) - \
+      (intptr_t)(((char*)((wxsPropertyContainer*)((Class*)1)))) )
 
 /** \brief Macro converting variable offset to value
  *
